% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{create_pop_cluster}
\alias{create_pop_cluster}
\title{Create point clusters for synthetic data testing}
\usage{
create_pop_cluster(x, y, dx, dy, n, crs = 32189)
}
\arguments{
\item{x}{Centre of distribution}

\item{y}{Centre of distribution}

\item{dx}{Standard deviation in x dimension}

\item{dy}{Standard deviation in y dimension}

\item{n}{Number of points}

\item{crs}{CRS, optional. Defaults to 32189 for Ottawa, ON.}
}
\value{
A simple feature collection of points.
}
\description{
Create point clusters for synthetic data testing
}
