% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{create_sli}
\alias{create_sli}
\title{Created single-link indicator from weighted links using maximum weights}
\usage{
create_sli(weighted_links, from_idcol, weights_col = "weight")
}
\arguments{
\item{weighted_links}{A `tbl_df` containing weighted links from one set of ids to another, such as created by `generate_weighted_links()`.}

\item{from_idcol}{Character. Name of column containing ids.}

\item{weights_col}{Character (default "weight"). Name of column containing weights.}
}
\value{
A `tbl_df` with two columns and one row for each unique value in input column `from_idcol`.
}
\description{
Created single-link indicator from weighted links using maximum weights
}
