% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{create_synthetic_population}
\alias{create_synthetic_population}
\title{Create synthetic population data for testing}
\usage{
create_synthetic_population(
  source_shape,
  num_clusters = 100,
  cluster_size = 2500,
  cluster_pop = 1000
)
}
\arguments{
\item{source_shape}{Input shape whose bounding box we'll use}

\item{num_clusters}{Number of clusters to create.}

\item{cluster_size}{SD of cluster size, in units of source_shape.}

\item{cluster_pop}{Population of clusters.}
}
\value{
Simple feature tibble with population clusters.
}
\description{
Create synthetic population data for testing
}
